pyproject.toml
[tool.poetry]
name = "s3wm"
version = "0.1.3"
description = "Simple Window manager."
authors = ["Pavel Kirilin <win10@list.ru>"]
readme = "README.md"
packages = [
    {include = "s3wm"},
    {include = "s3wm_core"}
]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
frozendict = "^1.2"
pydantic = "^1.7.3"
python-xlib = "^0.29"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
flake8 = "^3.8.4"
black = "^20.8b1"
mypy = "^0.800"
isort = "^5.7.0"
pre-commit = "^2.10.1"
yesqa = "^1.2.2"
autoflake = "^1.4"
wemake-python-styleguide = "^0.15.1"

[tool.poetry.scripts]
s3wm = "s3wm.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
